Created new Response component
